K Nearest Neighbours Trainer

Trains a k nearest neighbour model using fast search algorithms KNN is a supervised learning algorithm which is used for both regression and classification problems.

KNNTrainer

Format

R6Class object.

Usage

For usage details see Methods, Arguments and Examples sections.

bst = KNNTrainer$new(k, prob, algorithm, type)
bst$fit_predict(train, test, y)
bst$get_predictions(type)

Methods

$new()

Initialise the instance of the vectorizer

$fit_predict()

trains the knn model and stores the test prediction

$get_predictions()

returns predictions

Arguments

k

number of neighbours to predict

prob

if probablity should be computed

algorithm

algorithm used to train the model, possible values are 'kd_tree','cover_tree','brute'

type

type of problem to solve i.e. regression or classification, possible values are 'reg' or 'class'

Examples

data("iris")

iris$Species <- as.integer(as.factor(iris$Species))

xtrain <- iris[1:100,]
xtest <- iris[101:150,]

bst <- KNNTrainer$new(k=3, prob=TRUE, type="class")
bst$fit_predict(xtrain, xtest, 'Species')
pred <- bst$get_predictions(type="raw")