cla_train

Training Dataset used for classification examples. This is classic titanic dataset used to predict if a passenger will survive or not in titanic ship disaster.

cla_train

Format

A data frame with 12 variables:

PassengerId

unique ID of the passenger

Survived

survival of the passenger

Pclass

ticket class

Name

name of the passenger

Sex

sex of the passenger

Age

Age of the passenger

SibSp

number of siblings/spouse aboard the titanic

Parch

number of parents / child aboard the titanic

Ticket

ticket number

Fare

passenger fare

Cabin

cabin number

Embarked

port of embarkation

For further details, see https://www.kaggle.com/c/titanic/data