K-Means Trainer

Trains am unsupervised K-Means algorithm. It borrows mini-batch k-means function from ClusterR r package which is written in c++, hence it is quite fast.

KMeansTrainer

Format

R6Class object.

Usage

For usage details see Methods, Arguments and Examples sections.

kmt = KMeansTrainer$new(n_estimators, max_features, max_depth, min_node_size, criterion,classification, class_weights, verbose, seed,always_split)
bst$fit(X_train, y_train)
prediction <- bst$predict(X_test)

Methods

$new()

Initialises an instance of k-means model

$fit()

fit model to an input train data

$predict()

returns cluster predictions for each row of given data

Arguments

params

for explanation on parameters, please refer to the documentation of MiniBatchKMeans function in clusterR package https://cran.r-project.org/web/packages/ClusterR/ClusterR.pdf

find_optimal

Used to find the optimal number of cluster during fit method. To use this, make sure the value for /codemax_cluster > 0.

Examples

data <- rbind(replicate(20, rnorm(1e5, 2)),
             replicate(20, rnorm(1e5, -1)),
             replicate(20, rnorm(1e5, 5)))
km_model <- KMeansTrainer$new(clusters=2, batch_size=30, max_clusters=6)
km_model$fit(data, find_optimal = FALSE)

#> Using 2 clusters to learn from data

predictions <- km_model$predict(data)